+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,7 +1,7 @@
 package edu.kis.vh.nursery;
 
-import edu.kis.vh.nursery.list.IntLinkedList;
-import edu.kis.vh.nursery.list.StackInterface;
+import edu.kis.vh.nursery.newStackImplementaion.IntLinkedList;
+import edu.kis.vh.nursery.newStackImplementaion.StackInterface;
 
 public class DefaultCountingOutRhymer {
