+++ src/main/java/edu/kis/vh/nursery/newStackImplementaion/StackInterface.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.newStackImplementaion;
 
 public interface StackInterface {
